web-controller/requirements.txt
Flask==2.1.2
Flask-AppBuilder==4.1.1
Flask-Mail==0.9.1
celery==5.2.7
redis==4.3.3
click==8.1.3
Pillow==9.1.1
opencv-python==4.5.5.64
imutils==0.5.4
python-dotenv==0.20.0
minio~=7.1.8
psycopg2-binary
flask==2.1.2
waitress==2.1.1